refactor: extract register-project and score-run use cases, dissolve private api imports
api now imports public entry points (register_project, score_completed_evidence);
DEFAULT_MAX_SUBAGENTS/DEFAULT_TIME_LIMIT lose their misleading underscore.
