+++ src/quodeq/services/base.py
@@ -7,9 +7,8 @@ from typing import Protocol, runtime_checkable
 
 from quodeq.core.types import JobSnapshot, ViolationSummary
 from quodeq.shared.constants import (  # noqa: F401 — re-export for backward compat
-    _DEFAULT_MAX_SUBAGENTS,
-    _DEFAULT_POOL_BUDGET,
-    _DEFAULT_TIME_LIMIT,
+    DEFAULT_MAX_SUBAGENTS,
+    DEFAULT_TIME_LIMIT,
 )
 
 
@@ -23,8 +22,8 @@ class EvaluationOptions:
     ai_model: str | None = None
     subagent_model: str | None = None
     verify_findings: bool = True
-    max_subagents: int = _DEFAULT_MAX_SUBAGENTS
-    time_limit: int = _DEFAULT_TIME_LIMIT
+    max_subagents: int = DEFAULT_MAX_SUBAGENTS
+    time_limit: int = DEFAULT_TIME_LIMIT
     clean_scan: bool = False
     per_dimension: bool = False
     branch: str | None = None
