+++ tools/import_baseline.txt
@@ -16,8 +16,8 @@ src/quodeq/services/_external_jobs.py:21:analysis
 src/quodeq/services/_violations_jsonl.py:12:analysis
 src/quodeq/services/_violations_stream.py:10:analysis
 src/quodeq/services/_violations_stream.py:11:analysis
-src/quodeq/services/evaluation_mixin.py:22:analysis
-src/quodeq/services/evaluation_mixin.py:540:analysis
+src/quodeq/services/evaluation_mixin.py:360:analysis
 src/quodeq/services/jobs.py:20:analysis
 src/quodeq/services/scan_progress.py:23:analysis
+src/quodeq/services/score_run.py:14:analysis
 src/quodeq/services/tooling_mixin.py:17:analysis
